fix: announcement 리팩토링

+++ src/main/java/com/cpu/cams/announcement/service/AnnouncementService.java
@@ -4,16 +4,19 @@ import com.cpu.cams.announcement.dto.request.AnnouncementRequest;
 import com.cpu.cams.announcement.dto.response.AnnouncementResponse;
 import com.cpu.cams.announcement.entity.Announcement;
 import com.cpu.cams.announcement.repository.AnnouncementRepository;
+import com.cpu.cams.member.dto.response.CustomUserDetails;
 import com.cpu.cams.member.entity.Member;
 import com.cpu.cams.member.entity.Role;
 import com.cpu.cams.member.repository.MemberRepository;
 import com.cpu.cams.member.service.MemberService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -24,15 +27,16 @@ public class AnnouncementService {
     private final MemberService memberService;
 
     // 공지사항 등록
-    public Long createAnnouncement(AnnouncementRequest announcementRequest, String username) {
+    public Long createAnnouncement(AnnouncementRequest announcementRequest, CustomUserDetails customUserDetails) {
 
-        Member findMember = memberService.findByUsername(username);
-        isAdmin(findMember);
+        // 관리자 체크
+        checkAdmin(customUserDetails);
+
+        Member findMember = memberService.findByUsername(customUserDetails.getUsername());
 
         Announcement announcement = Announcement.create(announcementRequest, findMember);
         return announcement.getId();
     }
-
 
     // 공지사항 전체 조회
     public Page<AnnouncementResponse> getAnnouncements(int page, int size) {
@@ -53,29 +57,30 @@ public class AnnouncementService {
     }
 
     // 공지사항 수정
-    public void updateAnnouncement(Long announcementId, AnnouncementRequest announcementRequest, String username) {
+    public void updateAnnouncement(Long announcementId, AnnouncementRequest announcementRequest, CustomUserDetails customUserDetails) {
 
-        Member findMember = memberService.findByUsername(username);
-        isAdmin(findMember);
+        checkAdmin(customUserDetails);
 
         Announcement announcement = findOne(announcementId);
         announcement.updateAnnouncement(announcementRequest);
     }
 
     // 공지사항 삭제
-    public void deleteAnnouncement(Long announcementId, String username) {
+    public void deleteAnnouncement(Long announcementId, CustomUserDetails customUserDetails) {
 
-        Member findMember = memberService.findByUsername(username);
-        isAdmin(findMember);
+        checkAdmin(customUserDetails);
 
         Announcement announcement = findOne(announcementId);
         announcementRepository.delete(announcement);
     }
 
     // 관리자인지 확인하는 로직
-    private void isAdmin(Member findMember) {
-        if(!findMember.getRole().equals(Role.ROLE_ADMIN)){
-            throw new RuntimeException("너 관리자 맞아?");
+    private void checkAdmin(CustomUserDetails customUserDetails) {
+        boolean isAdmin = customUserDetails.getAuthorities().stream()
+                .anyMatch(a -> "ROLE_ADMIN".equals(a.getAuthority())); // 또는 Role.ROLE_ADMIN.name()
+
+        if (!isAdmin) {
+            throw new RuntimeException("너 관리자 아니구나?");
         }
     }
     
@@ -83,4 +88,5 @@ public class AnnouncementService {
     private Announcement findOne(Long announcementId){
         return announcementRepository.findById(announcementId).orElseThrow(() -> new RuntimeException("공지 없음"));
     }
+
 }
